Major refactor to group functionlaity and start making testable.
