+++ client-py/casperlabs_client/consts.py
@@ -1,5 +1,10 @@
+DEFAULT_HOST = "localhost"
+DEFAULT_PORT = 40401
+DEFAULT_INTERNAL_PORT = 40402
 STATUS_CHECK_DELAY = 0.5
 STATUS_TIMEOUT = 180  # 3 minutes
 DEFAULT_PAYMENT_AMOUNT = 10000000
-
+VISUALIZE_DAG_STREAM_DELAY = 5
 SUPPORTED_KEY_ALGORITHMS = ("ed25519", "bn2254")
+
+BUNDLED_TRANSFER_WASM: str = "transfer_to_account_u512.wasm"
